Add SSE4 support for 32-bit x86
Port the amd64 SSE4 support code to 32-bit x86.

This is mostly a straight-forward manual conversion, with most of the
actually new code being in some of the newly implemented Iops in
host_x86_isel.c and in the top-level block in disInstr_X86_WRK().

All FIXME comments and similar are from the original code.